requirements.txt
Django==4.2
gunicorn
whitenoise
tensorflow-cpu==2.15.0
numpy
opencv-python-headless
pandas
scikit-learn
h5py
pillow
django-cors-headers